requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
enum34==1.1.10
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.14.1
jupyter-client==6.1.5
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==0.24.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prompt-toolkit==2.0.10
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-model-optimization==0.3.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
